requirements.txt
# Python >=3.13,<3.14
numpy==2.3.2
rasterio==1.4.3
Pillow==11.3.0
tqdm==4.67.1
